requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
CacheControl==0.14.4
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
fastapi==0.121.2
firebase_admin==7.1.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.7.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
proto-plus==1.26.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
psycopg[binary]==3.1.19
